+++ CHANGELOG.md
@@ -260,6 +260,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Support for the analytics component
 - Function builder
+- Query builder
 - Solarium\Component\FacetSet::setMatches()
 - Solarium\Component\FacetSet::setExcludeTerms()
 - Solarium\Component\Facet\Field::setMatches()